actualizacion

+++ pysinergia/exportadores/exportador_word.py
@@ -14,7 +14,15 @@ class ExportadorWord(_I_Exportador):
 
     def generar(mi, contenido:str, destino:str=''):
         import subprocess, os, io
-        ruta_pandoc = os.path.join(os.path.abspath(_Constantes.DIR_LIB_PANDOC),'pandoc')
+        dir_pandoc = os.path.normpath(os.path.abspath(_Constantes.DIR_LIB_PANDOC))
+        """
+        import pandoc
+        os.environ["PATH"] = dir_pandoc + os.pathsep + os.getenv("PATH")
+        config = pandoc.configure(path=dir_pandoc, version="3.2.0", read=True)
+        doc = pandoc.read(contenido, format='html')
+        res = pandoc.write(doc=doc, format='docx', file=destino)
+        """
+        ruta_pandoc = os.path.join(dir_pandoc,'pandoc')
         ruta_temp = str(mi.opciones.get('ruta_temp', ''))
         dir_temp = f'{ruta_temp}/archivos'
         idioma = mi.opciones.get('idioma')
@@ -29,6 +37,7 @@ class ExportadorWord(_I_Exportador):
             docx_io = io.BytesIO(docx_bytes)
         os.remove(ruta_html)
         if not destino:
-            os.remove(salida_docx)
+            if os.path.exists(salida_docx):
+                os.remove(salida_docx)
         return docx_io
